+++ zhihu_oauth/__init__.py
@@ -1,16 +1,14 @@
 # coding=utf-8
 
-from __future__ import unicode_literals
-
 from .client import ZhihuClient
 from .zhcls import (
-    Answer, Article, Comment, Column, Collection, People, Question, Topic,
+    Activity, ActType, Answer, Article, Comment, Column, Collection, People, Question, Topic,
     ANONYMOUS
 )
 from .exception import NeedCaptchaException, UnexpectedResponseException
 
-__all__ = ['ZhihuClient', 'ANONYMOUS', 'Article', 'Answer', 'Collection',
-           'Column', 'Comment', 'People', 'Question', 'Topic',
-           'NeedCaptchaException', 'UnexpectedResponseException']
+__all__ = ['ZhihuClient', 'ANONYMOUS', 'Activity', 'ActType', 'Article',
+           'Answer', 'Collection', 'Column', 'Comment', 'People', 'Question',
+           'Topic', 'NeedCaptchaException', 'UnexpectedResponseException']
 
 __version__ = '0.0.3'